# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticatedUser": {
        "description": "Who just signed in — identity only.\n\nNo roles and no permissions: the effective permission union is resolved by\nF031's dependency and served by ``/auth/me``. Login answers \"here is your\nsession\", not \"here is everything about you\"; a second round trip after\nsign-in is the honest cost of keeping one definition of the access set.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "must_change_password"
        ],
        "title": "AuthenticatedUser",
        "type": "object"
      },
      "ChangePasswordRequest": {
        "description": "Self-service credential change (F030). Both fields are shape-bounded\nonly — see the module docstring for where the real policy lives.",
        "properties": {
          "current_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "required": [
          "user"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "MeResponse": {
        "description": "``/auth/me`` — identity plus the effective access set (F031).\n\nExtends the login response rather than restating it: the identity fields\nare the same four, and one definition of \"who is this\" beats two that\nalmost match. ``roles`` are names (the UI displays them; ids belong to the\nadmin editor, F033), and ``permissions`` is the **expanded union** — a\nsuperuser sees every code's name, never a wildcard, so the frontend checks\nset membership and never special-cases a flag. Both lists are sorted, so\nresponses are deterministic and diffable.\n\n``is_active``/``is_deleted`` are deliberately absent: a session belonging\nto a disabled account does not resolve at all (F029), so fields that could\nonly ever read \"true\" would be decoration. ``is_superuser`` is served —\nthe SPA's access model carries an explicit super-admin flag\n(``src/config/access.ts``, ARCHITECTURE §6's \"explicit super-admin\nhandling\"), and one truthful boolean beats a client that hardcodes\n``false`` and re-derives authority from grant-list length.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_superuser": {
            "title": "Is Superuser",
            "type": "boolean"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "must_change_password",
          "phone",
          "is_superuser",
          "roles",
          "permissions"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Application Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Change the caller's own password — the forced first-login flow and\nProfile > Security are the same request.\n\n204 with a **fresh cookie pair**: the session that asked is rotated (a\nnew ID, same family), every *other* session of the account is revoked\n(``password_change``), and ``must_change_password`` clears — one commit\nin the service. Refusals are field-addressable 422s (wrong current\npassword, policy violations, new equals current); a wrong current\npassword is deliberately *not* a 401, which this API reserves for \"your\nsession is over\" — a typo must not sign the user out. The current\npassword is itself throttled (its own bucket, 429 with ``Retry-After``),\nbecause a stolen session must not become a password oracle.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No usable session was presented."
          },
          "403": {
            "description": "CSRF check failed (see docs/ARCHITECTURE.md §3)."
          },
          "422": {
            "description": "Field-addressable (loc = body/current_password or body/new_password); no input value is echoed."
          },
          "429": {
            "description": "Too many failed current-password attempts; see Retry-After."
          }
        },
        "summary": "Change the signed-in user's password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid credentials (never says which part)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Login throttled; see Retry-After."
          }
        },
        "summary": "Sign in and start a session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "End the session this request presented, and clear both cookies.\n\nIdempotent: no cookie, a junk cookie, an already-ended session — the\nanswer is the same 204 and the cookies still leave the browser. The\nresolution that produced ``context`` is the full one, so a *rotated*\nID presented here trips the same family revocation as anywhere else; a\nstale tab is not a loophole, and neither is a thief.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "CSRF check failed (see docs/ARCHITECTURE.md §3)."
          }
        },
        "summary": "End the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "description": "Revoke every live session of the caller's account, this one included.\n\nDistinct from logout in the one way that matters: it acts under a session\n(401 without one), so it can revoke \"all *my other* sessions\" — the\nresponse to \"someone may have my credentials\". One bulk UPDATE; the\nfamily link is irrelevant here, because every family dies.",
        "operationId": "logout_all_api_v1_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No usable session was presented."
          },
          "403": {
            "description": "CSRF check failed (see docs/ARCHITECTURE.md §3)."
          }
        },
        "summary": "End every session of the signed-in user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Who the caller is, and what the database says they may do — right now.\n\nThe answer is computed from this request's resolution of the session,\nroles and permissions (F029 re-loads the graph every request), so a role\nchange made a second ago is already reflected and a revoked one is gone\n(BP-6.3f: no cache to invalidate). Deliberately reachable during a forced\npassword change — this is how the SPA learns the flag is set and where to\nroute — which is also why it does not ride the gated ``current_session``.\n\nThe permission list is the expanded union, superusers included (every\ncode, never a wildcard): the frontend checks set membership and never\nspecial-cases a flag (ARCHITECTURE §6 layer 3 — UX mirroring the server's\nreal answer).",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No usable session was presented."
          }
        },
        "summary": "The signed-in user and their effective permissions",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}